+++ CHANGELOG.md
@@ -4,6 +4,7 @@ CHANGELOG
 - **Unreleased**
   - [View Diff](https://github.com/westonganger/sexy_form.rb/compare/v0.9.0...master)
   - [#2](https://github.com/westonganger/sexy_form.rb/pulls/2) - Remove validations on field `:type` option
+  - Fix bug when field `:errors` is provided but is empty
 
 - **0.9.0** - February 15, 2019
   - Gem Initial Release